chore: move validators to validator file
